requirements.txt
asgiref==3.7.2
Django==4.2.8
django-filter==23.5
djangorestframework==3.14.0
importlib-metadata==7.0.0
Markdown==3.5.1
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
zipp==3.17.0
